Improve support for common Foundation macros
Previously code like the following would break this grammar:

```
- (void)sceneWillDeactivate:(NSNotification *)notification API_AVAILABLE(ios(13.0)) {
  ...
}
```

Since the `API_AVAILABLE(` would be treated actually as a return type
instead of a macro/function call. This was fixed with a lookbehind
assertion for the return-type tokens.

In addition, I've tried to improve support for common macros that can
contain arguments (supporting C constructs as parameters, Xcode
does something similar as well), but not sure how properly this works,
especially inside of VS Code.